+++ chakracore-cxx/lib/Backend/BailOut.cpp
@@ -13,6 +13,7 @@
 #endif
 #include "Language/JavascriptFunctionArgIndex.h"
 #include "LazyBailOutRecord.h"
+#include "Common/InlinedFrameLayout.h"
 
 extern const IRType RegTypes[RegNumCount];
