+++ chakracore-cxx/lib/Backend/InlineeFrameInfo.h
@@ -4,6 +4,7 @@
 //-------------------------------------------------------------------------------------------------------
 
 #pragma once
+#include "Common/InlinedFrameLayout.h"
 
 class Value;
